refactor: extract useRoutines hook
Moves all routines state, the day-rollover effect, and routines
handlers out of App.jsx into a dedicated useRoutines hook.

Hook accepts { currentTime, onboardingProgress, setOnboardingProgress }
and returns all 13 state values plus openRoutinesDashboard,
addRoutineChip, deleteRoutineChip, toggleRoutineChipSelection,
and handleRoutinesDone.

routinesEnabled is co-located with routines state as specified in
the refactoring plan.
